% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_aic}
\alias{fit_aic}
\title{Internal utility: compute AIC for a given variable set}
\usage{
fit_aic(vars_idx, df, var_names, family)
}
\description{
Internal utility: compute AIC for a given variable set
}
\keyword{internal}
